fix: seperating util functions into seperate file and simplifying condition to generate result
