+++ src/App.tsx
@@ -4,7 +4,7 @@ import ResultTable from "./components/ResultTable/ResultTable";
 import { ChangeEvent, useState } from "react";
 import Header from './components/Header/Header';
 import { Input } from "./components/UserInput/UserInput";
-import { AnnualData } from "./components/ResultTable/ResultTable";
+import { getResult } from "./utils/computeResult";
 
 export const InputEnums = {
   InitialInvestment: "initialInvestment",
@@ -14,40 +14,13 @@ export const InputEnums = {
 } as const
 export type InputEnumType = typeof InputEnums[keyof typeof InputEnums]
 
-function getResult(initialInvestment:number,annualInvestment:number,expectedReturn:number,duration:number){
-  const annualData:AnnualData = [];
-  let investmentValue = initialInvestment;
-
-  for (let i = 0; i < duration; i++) {
-    const interestEarnedInYear = investmentValue * (expectedReturn / 100);
-    investmentValue += interestEarnedInYear + annualInvestment;
-    if(i === 0){
-      annualData.push({
-        year: i + 1,
-        investmentValue: Math.ceil(investmentValue), 
-        interest: Math.ceil(interestEarnedInYear), 
-        totalInterest: Math.ceil(interestEarnedInYear), 
-        investedCapital: Math.ceil(initialInvestment+annualInvestment),
-      });
-    }else{
-      annualData.push({
-        year: i + 1,
-        investmentValue: Math.ceil(investmentValue), 
-        interest: Math.ceil(interestEarnedInYear), 
-        totalInterest: Math.ceil(annualData[i-1].interest + interestEarnedInYear), 
-        investedCapital: Math.ceil(annualData[i-1].investedCapital + annualInvestment),
-      });
-    }
-  }
-
-  return annualData;
-}
-
 function App() {
   const [initialInvestment, setInitialInvestment] = useState<Input>(undefined);
   const [annualInvestment, setAnnualInvestment] = useState<Input>(undefined);
   const [expectedReturn, setExpectedReturn] = useState<Input>(undefined);
   const [duration, setDuration] = useState<Input>(undefined);
+  const shouldGenerateResult = initialInvestment && annualInvestment && expectedReturn && duration
+  
   function handleUserInputchange(e: ChangeEvent<HTMLInputElement>,name:InputEnumType) {
     switch (name.toString()) {
       case InputEnums.InitialInvestment: {
@@ -72,7 +45,7 @@ function App() {
     <>
       <Header/>
       <UserInput initialInvestment={initialInvestment} annualInvestment={annualInvestment} expectedReturn={expectedReturn} duration={duration} inputChange={handleUserInputchange}/>
-      {initialInvestment && annualInvestment && expectedReturn && duration && (
+      {shouldGenerateResult && (
         <ResultTable resultData={getResult(initialInvestment,annualInvestment,expectedReturn,duration)}/>
       )}
     </>
